update tests in posix
git-svn-id: https://svn.code.sf.net/p/opensek/code/trunk@550 73ee5bea-b2a1-4b4e-8992-c3b1447581ab

+++ FreeOSEK/Os/src/posix/Os_Internal_Arch.c
@@ -163,7 +163,7 @@ void PosixInterruptHandler(int status)
 				printf("other error\n");
 				break;
 		}
-		printf("Error by reading the Message Queue, returned value: %d, error number: %d\n",mq_ret,errno);
+		printf("Error by reading the Message Queue, returned value: %d, error number: %d\n",(int)mq_ret,errno);
 	}
 
 	if (mq_notify(MessageQueue, &SignalEvent) == -1)
